chore: refactor into modern packaging style
