+++ .codecov.yml
@@ -1,0 +1,19 @@
+comment:
+  layout: "reach, diff, flags, files"
+  behavior: new
+  require_changes: true  # if true: only post the comment if coverage changes
+  require_base: no        # [yes :: must have a base report to post]
+  require_head: yes       # [yes :: must have a head report to post]
+
+codecov:
+  notify:
+    require_ci_to_pass: yes
+
+coverage:
+  status:
+    project: off
+    patch: off
+
+ignore:
+  - "./tests/"
+  - "./pyproject.toml"
