+++ CHANGELOG.md
@@ -1,0 +1,29 @@
+# CHANGELOG
+
+# Upcoming
+
+Added events table support.
+
+Added ability to specify additional metadata to write into the BIDS dataset.
+
+Debugged ecephys structure against BEP032 BIDS schema current to 6/26/2025.
+
+Refactored source code to modern Python packaging standards.
+
+
+
+# v0.0.3
+
+fixed sanitization syntax, added more robust tests
+
+
+
+# v0.0.2
+
+Make tests ignore order of files
+
+
+
+# v0.0.1
+
+Update proper file names
