--- setup.cfg
@@ -1,49 +1,0 @@
-[metadata]
-name = nwb2bids
-version = 0.0.1
-url = https://github.com/con/nwb2bids/
-author = nwb2bids developers
-description = Reorganize NWB files into a BIDS directory layout.
-long_description = file:README.md
-long_description_content_type = text/markdown; charset=UTF-8; variant=GFM
-license = MIT
-classifiers =
-    Development Status :: 4 - Beta
-    Intended Audience :: Science/Research
-    Topic :: Scientific/Engineering :: Information Analysis
-    License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3.12
-
-[options]
-python_requires = >=3.8
-install_requires =
-    numpy
-    pynwb
-    clize
-    pydantic
-packages = find:
-include_package_data = false
-zip_safe = false
-
-[options.extras_require]
-tests =
-    pytest
-
-[tool:pytest]
-log_cli = true
-
-[codespell]
-# Ref: https://github.com/codespell-project/codespell#using-a-config-file
-skip = .git*
-check-hidden = true
-# ignore-regex =
-# ignore-words-list =
-
-
-[options.entry_points]
-console_scripts =
-    nwb2bids = nwb2bids.cli:main
