+++ src/nwb2bids/_base/__init__.py
@@ -1,0 +1,1 @@
+from ._convert_nwb_dataset import convert_nwb_dataset
