+++ src/nwb2bids/_base/_convert_nwb_dataset.py
@@ -1,0 +1,263 @@
+import json
+import os
+import pathlib
+import shutil
+
+import pydantic
+import pynwb
+
+from ._events import _get_events_metadata, _get_events_table
+from ._utils import _drop_false_keys, _extract_metadata, _unique_list_of_dicts, _write_tsv
+from ..schemas import AdditionalMetadata
+
+
+@pydantic.validate_call
+def convert_nwb_dataset(
+    *,
+    nwb_directory: pydantic.DirectoryPath,
+    bids_directory: str | pathlib.Path,
+    copy: bool = True,
+    additional_metadata_file_path: pydantic.FilePath | None = None,
+) -> None:
+    """
+    Convert a directory of NWB files to BIDS format.
+
+    Parameters
+    ----------
+    nwb_directory : directory path
+        The path to the directory containing NWB files.
+    bids_directory : directory path
+        The path to the directory where the BIDS dataset will be created.
+    copy : bool
+        Whether to copy the NWB files into the BIDS dataset directory.
+        The default is True.
+    additional_metadata_file_path : file path, optional
+        The path to a JSON file containing additional metadata to be included in the BIDS dataset.
+        If not provided, the function will also look for a file named "additional_metadata.json" in the NWB directory.
+    """
+    bids_directory = pathlib.Path(bids_directory)
+    bids_directory.mkdir(exist_ok=True)
+    additional_metadata_file_path = (
+        secondary_path
+        if additional_metadata_file_path is None
+        and (secondary_path := pathlib.Path(nwb_directory) / "additional_metadata.json").exists()
+        else additional_metadata_file_path
+    )
+
+    additional_metadata_model = None
+    all_metadata = {}
+    if additional_metadata_file_path is not None:
+        with additional_metadata_file_path.open(mode="r") as file_stream:
+            additional_metadata_dict = json.load(fp=file_stream)
+        additional_metadata_model = AdditionalMetadata(**additional_metadata_dict)
+
+        # Top-level fields (required for BIDS)
+        # Possible that DANDI itself should be primarily responsible for modifying certain things at time of publication
+        dataset_description_file_path = os.path.join(bids_directory, "dataset_description.json")
+        with open(file=dataset_description_file_path, mode="w") as file_stream:
+            file_stream.write(additional_metadata_model.model_dump_json(indent=4))
+
+    # Fields within NWB files
+    nwb_files = list(pathlib.Path(nwb_directory).rglob("*.[nN][wW][bB]"))
+
+    for nwb_file in nwb_files:
+        metadata = _extract_metadata(nwb_file)
+        all_metadata[nwb_file] = _extract_metadata(nwb_file)
+
+    os.makedirs(bids_directory, exist_ok=True)
+
+    subjects = _unique_list_of_dicts([x["subject"] for x in all_metadata.values()])
+
+    subjects = _drop_false_keys(subjects)
+
+    subjects_file_path = os.path.join(bids_directory, "participants.tsv")
+    # BIDS validation enforces column order
+    # Follow-up TODO: make keys dynamic based on availability
+    # Follow-up TODO: generalize to more subjects
+    possible_subject_fields = ["participant_id", "species", "strain", "sex"]
+    subject_fields = [field for field in possible_subject_fields if subjects[0].get(field, None) is not None]
+    subject_header = "\t".join(subject_fields)
+    subject_lines = [f"{subject_header}\n"]
+    for subject in subjects:
+        line = "\t".join(subject[field] for field in subject_fields)
+        subject_lines.append(f"{line}\n")
+
+    # TSV writer below is hard to control header order - TSV is not hard to write directly, so just do it here...
+    with open(file=subjects_file_path, mode="w") as file_stream:
+        file_stream.writelines(subject_lines)
+
+    # create participants JSON
+    default_subjects_json = {
+        "subject_id": {"Description": "Unique identifier of the subject"},
+        "species": {"Description": "The binomial species name from the NCBI Taxonomy"},
+        "strain": {"Description": "Identifier of the strain"},
+        "birthdate": {"Description": "Day of birth of the participant in ISO8601 format"},
+        "age": {
+            "Description": "Age of the participant at time of recording",
+            "Units": "days",
+        },
+        "sex": {"Description": "Sex of participant"},
+    }
+
+    subjects_json = {k: v for k, v in default_subjects_json.items() if k in subject_fields}
+    with open(os.path.join(bids_directory, "participants.json"), "w") as json_file:
+        json.dump(subjects_json, json_file, indent=4)
+
+    # sessions
+
+    default_session_json = {
+        "session_quality": {
+            "LongName": "General quality of the session",
+            "Description": "Quality of the session",
+            "Levels": {
+                "Bad": "Bad quality, should not be considered for further analysis",
+                "ok": "Ok quality, can be considered for further analysis with care",
+                "good": "Good quality, should be used for analysis",
+                "Excellent": "Excellent quality, extraordinarily good session",
+            },
+        },
+        "data_quality": {
+            "LongName": "Quality of the recorded signals",
+            "Description": "Quality of the recorded signals",
+            "Levels": {
+                "Bad": "Bad quality, should not be considered for further analysis",
+                "ok": "Ok quality, can be considered for further analysis with care",
+                "good": "Good quality, should be used for analysis",
+                "Excellent": "Excellent quality, extraordinarily good session",
+            },
+        },
+        "number_of_trials": {
+            "LongName": "Number of trials in this session",
+            "Description": "Count of attempted trials in the session (integer)",
+        },
+        "comment": {
+            "LongName": "General comments",
+            "Description": "General comments by the experimenter on the session",
+        },
+    }
+
+    for subject in subjects:
+        participant_id = subject["participant_id"]
+
+        os.makedirs(os.path.join(bids_directory, participant_id), exist_ok=True)
+
+        for metadata in all_metadata.values():
+            sessions = [x["session"] for x in all_metadata.values() if x["subject"]["participant_id"] == participant_id]
+
+            sessions = _drop_false_keys(sessions)
+
+            sessions_file_path = os.path.join(bids_directory, participant_id, f"{participant_id}_sessions.tsv")
+            sessions_keys = _write_tsv(sessions, sessions_file_path)
+            sessions_json = {k: v for k, v in default_session_json.items() if k in sessions_keys}
+
+            with open(
+                os.path.join(bids_directory, participant_id, f"{participant_id}_sessions.json"),
+                "w",
+            ) as json_file:
+                json.dump(sessions_json, json_file, indent=4)
+
+    # electrodes, probes, and channels
+
+    for nwbfile_path, metadata in all_metadata.items():
+        participant_id = metadata["subject"]["participant_id"]
+        session_id = metadata["session"]["session_id"] or ""  # Follow-up TODO: cleanup the missing session ID case
+
+        print(participant_id, session_id)
+        if session_id:
+            os.makedirs(os.path.join(bids_directory, participant_id, session_id), exist_ok=True)
+        else:
+            os.makedirs(os.path.join(bids_directory, participant_id), exist_ok=True)
+        # Ephys might need to be dynamically selected, nwb can also be ieeg.
+        os.makedirs(os.path.join(bids_directory, participant_id, session_id, "ephys"), exist_ok=True)
+
+        # Temporary hack to get this to obey validation
+        file_prefix = (
+            f"{metadata['subject']['participant_id']}_{metadata['session']['session_id']}"
+            if metadata["session"].get("session_id", None) is not None
+            else f"{metadata['subject']['participant_id']}"
+        )
+        for var in ("electrodes", "probes", "channels"):
+            var_metadata = metadata[var]
+            # var_metadata = drop_false_keys(var_metadata)
+            var_metadata_file_path = os.path.join(
+                bids_directory,
+                participant_id,
+                session_id,
+                "ephys",
+                f"{file_prefix}_{var}.tsv",
+            )
+            _write_tsv(var_metadata, var_metadata_file_path)
+
+        # Write events.tsv file
+        with pynwb.NWBHDF5IO(path=nwbfile_path, mode="r") as file_stream:
+            nwbfile = file_stream.read()
+
+            nwb_events_table = _get_events_table(nwbfile=nwbfile)
+
+        if nwb_events_table is not None:
+            # Collapse 'start_time' and 'stop_time' columns into 'onset' and 'duration' columns
+            bids_event_table = nwb_events_table.copy()
+            bids_event_table["duration"] = bids_event_table["stop_time"] - bids_event_table["start_time"]
+            bids_event_table = bids_event_table.rename(columns={"start_time": "onset"})
+            bids_event_table = bids_event_table.drop(columns=["stop_time"])
+            bids_event_table = bids_event_table.sort_values(
+                by=["onset", "duration"], ascending=[True, False]
+            ).reset_index(drop=True)
+
+            # BIDS Validator is strict regarding column order
+            required_column_order = ["onset", "duration", "nwb_table"]
+            all_columns = list(bids_event_table.columns)
+            final_column_order = required_column_order + [
+                column for column in all_columns if column not in required_column_order
+            ]
+
+            session_events_table_file_path = (
+                bids_directory / participant_id / session_id / "ephys" / f"{file_prefix}_events.tsv"
+            )
+            bids_event_table.to_csv(
+                path_or_buf=session_events_table_file_path,
+                sep="\t",
+                index=False,
+                columns=final_column_order,
+            )
+
+            # Write events.json file
+            bids_event_metadata = _get_events_metadata(nwbfile=nwbfile)
+            additional_metadata_events = (
+                additional_metadata_model.events.model_dump()
+                if additional_metadata_model is not None
+                else None or dict()
+            )
+            for key, value in additional_metadata_events.items():
+                if key not in bids_event_metadata:
+                    bids_event_metadata[key] = value
+                else:
+                    bids_event_metadata[key].update(value)
+
+            session_events_metadata_file_path = (
+                bids_directory / participant_id / session_id / "ephys" / f"{file_prefix}_events.json"
+            )
+            with session_events_metadata_file_path.open(mode="w") as file_stream:
+                json.dump(obj=bids_event_metadata, fp=file_stream, indent=4)
+
+        # Follow-up TODO: check events.json files, see if all are the same
+        # and if so remove the duplicates and move to outer level
+
+        # Rename and/or copy NWB file
+
+        bids_path = os.path.join(bids_directory, participant_id)
+        if metadata["session"]["session_id"]:
+            bids_path = os.path.join(bids_path, session_id)
+        bids_path = os.path.join(bids_directory, participant_id)
+
+        bids_path = os.path.join(
+            bids_directory,
+            metadata["subject"]["participant_id"],
+            session_id,
+            "ephys",
+            f"{file_prefix}_ephys.nwb",
+        )
+        if copy is True:
+            shutil.copyfile(src=nwb_file, dst=bids_path)
+        else:
+            open(bids_path, "a").close()
