+++ src/nwb2bids/_hidden_top_level_imports.py
@@ -1,0 +1,9 @@
+"""
+Including these directly within the top-level `__init__.py` makes them visible to autocompletion.
+
+But we only want the imports to trigger, not for them to actually be exposed.
+"""
+
+from .testing import assert_subdirectory_structure
+
+_hide = True
