+++ src/nwb2bids/schemas/__init__.py
