+++ src/nwb2bids/schemas/_additional_metadata_schema.py
