+++ src/nwb2bids/testing/__init__.py
@@ -1,0 +1,5 @@
+from ._assert_subdirectory_structure import assert_subdirectory_structure
+
+__all__ = [
+    "assert_subdirectory_structure",
+]
