+++ test/conftest.py
@@ -1,9 +1,9 @@
 import json
-
-import pynwb
-import numpy
-import pytest
 from datetime import datetime
+
+import numpy
+import pynwb
+import pytest
 from dateutil import tz
 
 
@@ -13,12 +13,8 @@ def _generate_mock_trials_table() -> pynwb.epoch.TimeIntervals:
 
     These are shorter than epochs and fit neatly within them in a qualitative manner.
     """
-    trials = pynwb.epoch.TimeIntervals(
-        name="trials", description="A mock trials table."
-    )
-    trials.add_column(
-        name="trial_condition", description="Extra information per trial."
-    )
+    trials = pynwb.epoch.TimeIntervals(name="trials", description="A mock trials table.")
+    trials.add_column(name="trial_condition", description="Extra information per trial.")
     trials.add_row(start_time=0.0, stop_time=1.0, trial_condition="A")
     trials.add_row(start_time=2.0, stop_time=3.0, trial_condition="B")
     trials.add_row(start_time=5.0, stop_time=5.5, trial_condition="C")
@@ -33,12 +29,8 @@ def _generate_mock_epochs_table() -> pynwb.epoch.TimeIntervals:
 
     These are longer than trials and encompass several of them in a qualitative manner.
     """
-    epochs = pynwb.epoch.TimeIntervals(
-        name="epochs", description="A mock epochs table."
-    )
-    epochs.add_column(
-        name="epoch_condition", description="Extra information per epoch."
-    )
+    epochs = pynwb.epoch.TimeIntervals(name="epochs", description="A mock epochs table.")
+    epochs.add_column(name="epoch_condition", description="Extra information per epoch.")
     epochs.add_row(start_time=0.0, stop_time=3.5, epoch_condition="AB")
     epochs.add_row(start_time=4.5, stop_time=6.0, epoch_condition="CD")
 
@@ -47,12 +39,8 @@ def _generate_mock_epochs_table() -> pynwb.epoch.TimeIntervals:
 
 def _generate_mock_time_intervals() -> pynwb.epoch.TimeIntervals:
     """Generate a mock time intervals table."""
-    time_intervals = pynwb.epoch.TimeIntervals(
-        name="mock_time_intervals", description="A mock time intervals table."
-    )
-    time_intervals.add_column(
-        name="tag", description="A tag assigned to each interval."
-    )
+    time_intervals = pynwb.epoch.TimeIntervals(name="mock_time_intervals", description="A mock time intervals table.")
+    time_intervals.add_column(name="tag", description="A tag assigned to each interval.")
     time_intervals.add_row(start_time=0.5, stop_time=0.75, tag="tag1")
     time_intervals.add_row(start_time=5.1, stop_time=5.23, tag="tag2")
 
@@ -67,18 +55,14 @@ def _generate_mock_time_intervals() -> pynwb.epoch.TimeIntervals:
 def nwb_testdata_multiple_events(
     tmp_path_factory,
 ):
-    from pynwb.testing.mock.file import mock_NWBFile
     from pynwb.file import Subject
+    from pynwb.testing.mock.file import mock_NWBFile
 
     nwbfile = mock_NWBFile(
-        session_start_time=datetime(
-            2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")
-        ),
+        session_start_time=datetime(2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")),
         session_id="20240309",
     )
-    time_series = pynwb.TimeSeries(
-        name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0
-    )
+    time_series = pynwb.TimeSeries(name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0)
     nwbfile.add_acquisition(time_series)
 
     subject = Subject(
@@ -107,18 +91,14 @@ def nwb_testdata_multiple_events(
 def nwb_testdata_trials_events(
     tmp_path_factory,
 ):
-    from pynwb.testing.mock.file import mock_NWBFile
     from pynwb.file import Subject
+    from pynwb.testing.mock.file import mock_NWBFile
 
     nwbfile = mock_NWBFile(
-        session_start_time=datetime(
-            2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")
-        ),
+        session_start_time=datetime(2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")),
         session_id="20240309",
     )
-    time_series = pynwb.TimeSeries(
-        name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0
-    )
+    time_series = pynwb.TimeSeries(name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0)
     nwbfile.add_acquisition(time_series)
 
     subject = Subject(
@@ -141,18 +121,14 @@ def nwb_testdata_trials_events(
 def nwb_testdata_epochs_events(
     tmp_path_factory,
 ):
-    from pynwb.testing.mock.file import mock_NWBFile
     from pynwb.file import Subject
+    from pynwb.testing.mock.file import mock_NWBFile
 
     nwbfile = mock_NWBFile(
-        session_start_time=datetime(
-            2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")
-        ),
+        session_start_time=datetime(2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")),
         session_id="20240309",
     )
-    time_series = pynwb.TimeSeries(
-        name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0
-    )
+    time_series = pynwb.TimeSeries(name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0)
     nwbfile.add_acquisition(time_series)
 
     subject = Subject(
@@ -175,18 +151,14 @@ def nwb_testdata_epochs_events(
 def nwb_testdata(
     tmp_path_factory,
 ):
-    from pynwb.testing.mock.file import mock_NWBFile
     from pynwb.file import Subject
+    from pynwb.testing.mock.file import mock_NWBFile
 
     nwbfile = mock_NWBFile(
-        session_start_time=datetime(
-            2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")
-        ),
+        session_start_time=datetime(2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")),
         session_id="20240309",
     )
-    time_series = pynwb.TimeSeries(
-        name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0
-    )
+    time_series = pynwb.TimeSeries(name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0)
     nwbfile.add_acquisition(time_series)
 
     subject = Subject(
@@ -202,19 +174,15 @@ def nwb_testdata(
 
 
 @pytest.fixture(scope="session")
-def nwb_testdata_nosessionid(tmp_path_factory):
-    from pynwb.testing.mock.file import mock_NWBFile
+def nwb_testdata_no_session_id(tmp_path_factory):
     from pynwb.file import Subject
+    from pynwb.testing.mock.file import mock_NWBFile
 
     nwbfile = mock_NWBFile(
-        session_start_time=datetime(
-            2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")
-        ),
+        session_start_time=datetime(2024, 3, 9, 22, 30, 3, tzinfo=tz.gettz("US/Eastern")),
         session_id="",
     )
-    time_series = pynwb.TimeSeries(
-        name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0
-    )
+    time_series = pynwb.TimeSeries(name="Test", data=numpy.array([], dtype="uint8"), unit="n.a.", rate=1.0)
     nwbfile.add_acquisition(time_series)
 
     subject = Subject(
@@ -244,9 +212,7 @@ def additional_metadata_fixture(
         }
     }
 
-    file_path = (
-        tmp_path_factory.mktemp("test_nwb2bids") / "test_additional_metadata.json"
-    )
+    file_path = tmp_path_factory.mktemp("test_nwb2bids") / "test_additional_metadata.json"
     with file_path.open(mode="w") as file_stream:
         json.dump(obj=additional_metdata_dictionary, fp=file_stream)
